+++ NEWS.md
@@ -1,7 +1,23 @@
 <!-- README.md is generated from README.Rmd. Please edit that file -->
 
-New Featurs in R6DS 1.1.0
-=========================
+New Features in R6DS 1.2.0
+==========================
+
+-   The efficiency of the classes has been improved.
+-   Bugs fixed.
+
+New Features in R6DS 1.1.1
+==========================
+
+-   Bugs fixed.
+-   A new vignette was added into the package.
+-   Every class has the new common method `toList`.
+-   Every class has the new common method `is_empty`.
+-   The `insert` function of the `RBST` class can take more arguments.
+-   RStack has `peek` and RQueue has `peekleft`.
+
+New Features in R6DS 1.1.0
+==========================
 
 -   All the functions
 -   Documentation including README, LICENSE and etc.